work dir v4: delete invalid armhf triplet pkgs (!1801)
Alpine's abuild has a arch_to_hostspec function, which translates an
architecture name ("x86_64") to a hostspec triplet
("x86_64-alpine-linux-musl"). We are using this function in some
APKBUILDs in pmaports.git. The triplet had been changed in [1] and
now we must rebuild all relevant packages, or we will run into weird
issues like [2].

As some of these packages are generated with "pmbootstrap aportgen",
we can't simply bump the pkgrel and force a rebuild. That would break
the upstream compatibility checks, which rely on pkgver and pkgrel
being the same as the originating packages from Alpine.

Create a new workdir version with a migration that zaps the chroots and
deletes the affected package versions from the cache. Before this gets
merged, I will rebuild the affected packages in the binary package
repository, so pmbootstrap will always use the fixed packages after the
migration.

[1] https://github.com/alpinelinux/abuild/pull/56
[2] https://gitlab.com/postmarketOS/pmaports/issues/295