kconfig check: check for FW_LOADER_USER_HELPER
The history of the option is follows:

Prior to kernel version 3.7, the kernel had no capability of finding
firmware itself so it had to call out to userspace to find the firmware
for it. This changed with kernel 3.7 where the kernel gained "direct
loading support" and where the userspace helper was only called as a
fallback. In kernel 3.9 this functionality was put behind the kconfig
option FW_LOADER_USER_HELPER which is what this patch checks to be
enabled.

Without this option, the udev rule present in postmarketos-base won't be
notified that firmware needs to be loaded and as such can't call
firmwareload.sh which loads firmware from /lib/firmware/postmarketos.